Separate dialogs into modules within gui
